Engineer

See merge request aruw/controls/aruw-mcb!879

+++ aruw-mcb-project/src/aruwsrc/control/bounded-subsystem/bounded_subsystem_interface.hpp
@@ -51,14 +51,14 @@ public:
     virtual bool homedAndBounded() const = 0;
 
     /**
-     * Returns the upper bound in motor encoder ticks.
+     * Returns the upper bound.
      */
-    virtual uint64_t getUpperBound() const = 0;
+    virtual float getUpperBound() const = 0;
 
     /**
-     * Returns the lower bound in motor encoder ticks.
+     * Returns the lower bound.
      */
-    virtual uint64_t getLowerBound() const = 0;
+    virtual float getLowerBound() const = 0;
 
 protected:
     /**
@@ -79,9 +79,9 @@ protected:
      */
     virtual void stopDuringHoming() = 0;
     /**
-     * Sets the given motor encoder position to be the "home" of the subsystem's motor.
+     * Sets the given position to be the "home" of the subsystem's motor.
      */
-    virtual void setHome(uint64_t encoderPosition) = 0;
+    virtual void setHome(float encoderPosition) = 0;
 };  // class HomeableSubsystemInterface
 }  // namespace aruwsrc::control
